// rs-frontend/src/app/model/equipmentRequestDto.ts
/**
 * RoomEquipment API
 * This API is used to provide availability to customize rooms and their equipment. <br><br> Only users with given authorities can use it: <br><ul> <li>ADMIN</li> <li>SUPER ADMIN</li> </ul>
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

/**
 * Contains information about given equipment item
 */
export interface EquipmentRequestDto { 
    /**
     * Equipment name
     */
    name?: string;
    /**
     * Detailed description of item
     */
    description?: string;
    /**
     * Number of room where given item is currently placed
     */
    roomNumber?: number;
}